On second thoughts, let's stick with malloc
We don't want to mess about zeroing memory whenever we reallocate; plus
there's no need to guarantee memory is zero'd. This way, calloc is just
as trivial to implement.

We also change the resizing amount to account for the overlap. Now we
only ever need to resize once when there's not enough space.